Revert "Add Resettable interface; refactor & move all ContextHolders"
This commit muddies the diff; will refactor on another branch

This reverts commit 4a9e2898c7112381dc0ac2031cb1f77a27bd8207.
