+++ backend/src/main/java/gov/cdc/usds/simplereport/api/ApiUserContextHolder.java
@@ -1,4 +1,4 @@
-package gov.cdc.usds.simplereport.api.context;
+package gov.cdc.usds.simplereport.api;
 
 import gov.cdc.usds.simplereport.db.model.ApiUser;
 import org.springframework.context.annotation.Scope;
@@ -8,22 +8,18 @@ import org.springframework.web.context.WebApplicationContext;
 
 @Repository
 @Scope(scopeName = WebApplicationContext.SCOPE_REQUEST, proxyMode = ScopedProxyMode.TARGET_CLASS)
-public class ApiUserContextHolder implements Resettable {
-  private ApiUser _apiUser = null;
+public class ApiUserContextHolder {
+  private ApiUser apiUser = null;
 
   public ApiUser getCurrentApiUser() {
-    return _apiUser;
+    return apiUser;
   }
 
   public boolean hasBeenPopulated() {
-    return _apiUser != null;
+    return apiUser != null;
   }
 
   public void setCurrentApiUser(ApiUser apiUser) {
-    this._apiUser = apiUser;
-  }
-
-  public void reset() {
-    _apiUser = null;
+    this.apiUser = apiUser;
   }
 }
