+++ backend/src/main/java/gov/cdc/usds/simplereport/api/CurrentAccountRequestContextHolder.java
@@ -1,4 +1,4 @@
-package gov.cdc.usds.simplereport.api.context;
+package gov.cdc.usds.simplereport.api;
 
 import org.springframework.context.annotation.Scope;
 import org.springframework.context.annotation.ScopedProxyMode;
@@ -7,14 +7,14 @@ import org.springframework.web.context.WebApplicationContext;
 
 @Repository
 @Scope(scopeName = WebApplicationContext.SCOPE_REQUEST, proxyMode = ScopedProxyMode.TARGET_CLASS)
-public class SmsWebhookContextHolder {
-  private boolean _isSmsWebhook = false;
+public class CurrentAccountRequestContextHolder {
+  private boolean _isAccountRequest = false;
 
-  public void setIsSmsWebhook(boolean status) {
-    this._isSmsWebhook = status;
+  public void setIsAccountRequest(boolean status) {
+    this._isAccountRequest = status;
   }
 
-  public boolean isSmsWebhook() {
-    return _isSmsWebhook;
+  public boolean isAccountRequest() {
+    return _isAccountRequest;
   }
 }
