--- backend/src/main/java/gov/cdc/usds/simplereport/api/context/Resettable.java
@@ -1,5 +1,0 @@
-package gov.cdc.usds.simplereport.api.context;
-
-public interface Resettable {
-  void reset();
-}
