+++ backend/src/main/java/gov/cdc/usds/simplereport/api/pxp/CurrentPatientContextHolder.java
@@ -1,4 +1,4 @@
-package gov.cdc.usds.simplereport.api.context;
+package gov.cdc.usds.simplereport.api.pxp;
 
 import gov.cdc.usds.simplereport.db.model.Organization;
 import gov.cdc.usds.simplereport.db.model.PatientLink;
