+++ backend/src/main/java/gov/cdc/usds/simplereport/config/authorization/UserAuthorizationVerifier.java
@@ -1,6 +1,6 @@
 package gov.cdc.usds.simplereport.config.authorization;
 
-import gov.cdc.usds.simplereport.api.context.CurrentAccountRequestContextHolder;
+import gov.cdc.usds.simplereport.api.CurrentAccountRequestContextHolder;
 import gov.cdc.usds.simplereport.api.model.errors.NonexistentQueueItemException;
 import gov.cdc.usds.simplereport.api.model.errors.NonexistentUserException;
 import gov.cdc.usds.simplereport.api.model.errors.UnidentifiedUserException;
