+++ backend/src/main/java/gov/cdc/usds/simplereport/logging/AuditLoggingAdvice.java
@@ -1,6 +1,6 @@
 package gov.cdc.usds.simplereport.logging;
 
-import gov.cdc.usds.simplereport.api.context.CurrentPatientContextHolder;
+import gov.cdc.usds.simplereport.api.pxp.CurrentPatientContextHolder;
 import gov.cdc.usds.simplereport.db.model.Organization;
 import gov.cdc.usds.simplereport.db.model.PatientLink;
 import gov.cdc.usds.simplereport.service.AuditService;
