+++ backend/src/main/java/gov/cdc/usds/simplereport/logging/RestAuditLogManager.java
@@ -1,7 +1,7 @@
 package gov.cdc.usds.simplereport.logging;
 
-import gov.cdc.usds.simplereport.api.context.CurrentPatientContextHolder;
-import gov.cdc.usds.simplereport.api.context.SmsWebhookContextHolder;
+import gov.cdc.usds.simplereport.api.SmsWebhookContextHolder;
+import gov.cdc.usds.simplereport.api.pxp.CurrentPatientContextHolder;
 import gov.cdc.usds.simplereport.db.model.Organization;
 import gov.cdc.usds.simplereport.db.model.PatientLink;
 import gov.cdc.usds.simplereport.service.AuditService;
