+++ backend/src/test/java/gov/cdc/usds/simplereport/api/pxp/SmsCallbackControllerTest.java
@@ -5,10 +5,9 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import gov.cdc.usds.simplereport.api.BaseFullStackTest;
+import gov.cdc.usds.simplereport.api.CurrentAccountRequestContextHolder;
 import gov.cdc.usds.simplereport.api.ResourceLinks;
-import gov.cdc.usds.simplereport.api.context.CurrentAccountRequestContextHolder;
-import gov.cdc.usds.simplereport.api.context.CurrentPatientContextHolder;
-import gov.cdc.usds.simplereport.api.context.SmsWebhookContextHolder;
+import gov.cdc.usds.simplereport.api.SmsWebhookContextHolder;
 import gov.cdc.usds.simplereport.db.model.Facility;
 import gov.cdc.usds.simplereport.db.model.Organization;
 import gov.cdc.usds.simplereport.db.model.PatientLink;
